refactor: rename details to impl
