+++ packages/nvmedit/src/lib/nvm500/impls/Bridge_6_6x.ts
@@ -1,18 +1,17 @@
 import { MAX_NODES, NUM_NODEMASK_BYTES } from "@zwave-js/core/safe";
 import { SUC_MAX_UPDATES } from "../../../consts";
-import type { NVM500Details } from "../shared";
+import type { NVM500Impl } from "../shared";
 import {
 	APPL_NODEPARM_MAX,
 	NVMEntryType,
 	type NVMLayout,
 	NVM_SERIALAPI_HOST_SIZE,
-	POWERLEVEL_CHANNELS,
 	RTC_TIMER_SIZE,
 	SUC_CONTROLLER_LIST_SIZE,
 	TOTAL_RTC_TIMER_MAX,
 } from "../shared";
 
-const NVM_Layout_Bridge_6_8x: NVMLayout = [
+const NVM_Layout_Bridge_6_6x: NVMLayout = [
 	{ name: "nvmTotalEnd", type: NVMEntryType.Word, count: 1 },
 	{ name: "nvmZWlibrarySize", type: NVMEntryType.NVMModuleSize, count: 1 },
 	{ name: "NVM_INTERNAL_RESERVED_1_far", type: NVMEntryType.Byte, count: 4 },
@@ -128,13 +127,6 @@ const NVM_Layout_Bridge_6_8x: NVMLayout = [
 		count: 1,
 	},
 	{
-		name: "NVM_SECURITY0_KEY_far",
-		type: NVMEntryType.Buffer,
-		size: 16,
-		count: 1,
-	},
-	{ name: "NVM_SYSTEM_STATE", type: NVMEntryType.Byte, count: 1 },
-	{
 		name: "nvmZWlibraryDescriptor",
 		type: NVMEntryType.NVMModuleDescriptor,
 		count: 1,
@@ -144,7 +136,7 @@ const NVM_Layout_Bridge_6_8x: NVMLayout = [
 		type: NVMEntryType.NVMModuleSize,
 		// The Bridge API saves an additional node mask for the virtual nodes in the
 		// previous module, so we can use this offset to distinguish between the two.
-		offset: 0x2fef,
+		offset: 0x2fde,
 		count: 1,
 	},
 	{ name: "EEOFFSET_MAGIC_far", type: NVMEntryType.Byte, count: 1 },
@@ -157,31 +149,6 @@ const NVM_Layout_Bridge_6_8x: NVMLayout = [
 	{
 		name: "EEOFFSET_WATCHDOG_STARTED_far",
 		type: NVMEntryType.Byte,
-		count: 1,
-	},
-	{
-		name: "EEOFFSET_POWERLEVEL_NORMAL_far",
-		type: NVMEntryType.Byte,
-		count: POWERLEVEL_CHANNELS,
-	},
-	{
-		name: "EEOFFSET_POWERLEVEL_LOW_far",
-		type: NVMEntryType.Byte,
-		count: POWERLEVEL_CHANNELS,
-	},
-	{
-		name: "EEOFFSET_MODULE_POWER_MODE_EXTINT_ENABLE_far",
-		type: NVMEntryType.Byte,
-		count: 1,
-	},
-	{
-		name: "EEOFFSET_MODULE_POWER_MODE_far",
-		type: NVMEntryType.Byte,
-		count: 1,
-	},
-	{
-		name: "EEOFFSET_MODULE_POWER_MODE_WUT_TIMEOUT_far",
-		type: NVMEntryType.DWord,
 		count: 1,
 	},
 	{
@@ -215,20 +182,9 @@ const NVM_Layout_Bridge_6_8x: NVMLayout = [
 	{ name: "nvmModuleSizeEndMarker", type: NVMEntryType.Word, count: 1 },
 ];
 
-export const Bridge_6_8x: NVM500Details = {
-	name: "Bridge 6.8x",
+export const Bridge_6_6x: NVM500Impl = {
+	name: "Bridge 6.6x",
 	library: "bridge",
-	protocolVersions: [
-		"6.01",
-		"6.02",
-		"6.03",
-		"6.04",
-		"6.05",
-		"6.06",
-		"6.07",
-		"6.08",
-		"6.09",
-		"6.10",
-	],
-	layout: NVM_Layout_Bridge_6_8x,
+	protocolVersions: ["4.33", "4.62"],
+	layout: NVM_Layout_Bridge_6_6x,
 };
