+++ packages/nvmedit/src/lib/nvm500/impls/Bridge_6_7x.ts
@@ -1,17 +1,18 @@
 import { MAX_NODES, NUM_NODEMASK_BYTES } from "@zwave-js/core/safe";
 import { SUC_MAX_UPDATES } from "../../../consts";
-import type { NVM500Details } from "../shared";
+import type { NVM500Impl } from "../shared";
 import {
 	APPL_NODEPARM_MAX,
 	NVMEntryType,
 	type NVMLayout,
 	NVM_SERIALAPI_HOST_SIZE,
+	POWERLEVEL_CHANNELS,
 	RTC_TIMER_SIZE,
 	SUC_CONTROLLER_LIST_SIZE,
 	TOTAL_RTC_TIMER_MAX,
 } from "../shared";
 
-const NVM_Layout_Static_6_6x: NVMLayout = [
+const NVM_Layout_Bridge_6_7x: NVMLayout = [
 	{ name: "nvmTotalEnd", type: NVMEntryType.Word, count: 1 },
 	{ name: "nvmZWlibrarySize", type: NVMEntryType.NVMModuleSize, count: 1 },
 	{ name: "NVM_INTERNAL_RESERVED_1_far", type: NVMEntryType.Byte, count: 4 },
@@ -99,6 +100,11 @@ const NVM_Layout_Static_6_6x: NVMLayout = [
 		count: 1,
 	},
 	{
+		name: "EX_NVM_BRIDGE_NODEPOOL_START_far",
+		type: NVMEntryType.NodeMask,
+		count: 1,
+	},
+	{
 		name: "EX_NVM_CONTROLLER_CONFIGURATION_far",
 		type: NVMEntryType.Byte,
 		count: 1,
@@ -122,6 +128,12 @@ const NVM_Layout_Static_6_6x: NVMLayout = [
 		count: 1,
 	},
 	{
+		name: "NVM_SECURITY0_KEY_far",
+		type: NVMEntryType.Buffer,
+		size: 16,
+		count: 1,
+	},
+	{
 		name: "nvmZWlibraryDescriptor",
 		type: NVMEntryType.NVMModuleDescriptor,
 		count: 1,
@@ -131,7 +143,7 @@ const NVM_Layout_Static_6_6x: NVMLayout = [
 		type: NVMEntryType.NVMModuleSize,
 		// The Bridge API saves an additional node mask for the virtual nodes in the
 		// previous module, so we can use this offset to distinguish between the two.
-		offset: 0x2fc1,
+		offset: 0x2fee,
 		count: 1,
 	},
 	{ name: "EEOFFSET_MAGIC_far", type: NVMEntryType.Byte, count: 1 },
@@ -145,6 +157,16 @@ const NVM_Layout_Static_6_6x: NVMLayout = [
 		name: "EEOFFSET_WATCHDOG_STARTED_far",
 		type: NVMEntryType.Byte,
 		count: 1,
+	},
+	{
+		name: "EEOFFSET_POWERLEVEL_NORMAL_far",
+		type: NVMEntryType.Byte,
+		count: POWERLEVEL_CHANNELS,
+	},
+	{
+		name: "EEOFFSET_POWERLEVEL_LOW_far",
+		type: NVMEntryType.Byte,
+		count: POWERLEVEL_CHANNELS,
 	},
 	{
 		name: "nvmApplicationDescriptor",
@@ -177,9 +199,9 @@ const NVM_Layout_Static_6_6x: NVMLayout = [
 	{ name: "nvmModuleSizeEndMarker", type: NVMEntryType.Word, count: 1 },
 ];
 
-export const Static_6_6x: NVM500Details = {
-	name: "Static 6.6x",
-	library: "static",
-	protocolVersions: ["4.33", "4.62"],
-	layout: NVM_Layout_Static_6_6x,
+export const Bridge_6_7x: NVM500Impl = {
+	name: "Bridge 6.7x",
+	library: "bridge",
+	protocolVersions: ["4.60", "4.61", "5.02", "5.03"],
+	layout: NVM_Layout_Bridge_6_7x,
 };
