+++ packages/nvmedit/src/lib/nvm500/impls/Static_6_6x.ts
@@ -1,6 +1,6 @@
 import { MAX_NODES, NUM_NODEMASK_BYTES } from "@zwave-js/core/safe";
 import { SUC_MAX_UPDATES } from "../../../consts";
-import type { NVM500Details } from "../shared";
+import type { NVM500Impl } from "../shared";
 import {
 	APPL_NODEPARM_MAX,
 	NVMEntryType,
@@ -11,7 +11,7 @@ import {
 	TOTAL_RTC_TIMER_MAX,
 } from "../shared";
 
-const NVM_Layout_Bridge_6_6x: NVMLayout = [
+const NVM_Layout_Static_6_6x: NVMLayout = [
 	{ name: "nvmTotalEnd", type: NVMEntryType.Word, count: 1 },
 	{ name: "nvmZWlibrarySize", type: NVMEntryType.NVMModuleSize, count: 1 },
 	{ name: "NVM_INTERNAL_RESERVED_1_far", type: NVMEntryType.Byte, count: 4 },
@@ -99,11 +99,6 @@ const NVM_Layout_Bridge_6_6x: NVMLayout = [
 		count: 1,
 	},
 	{
-		name: "EX_NVM_BRIDGE_NODEPOOL_START_far",
-		type: NVMEntryType.NodeMask,
-		count: 1,
-	},
-	{
 		name: "EX_NVM_CONTROLLER_CONFIGURATION_far",
 		type: NVMEntryType.Byte,
 		count: 1,
@@ -136,7 +131,7 @@ const NVM_Layout_Bridge_6_6x: NVMLayout = [
 		type: NVMEntryType.NVMModuleSize,
 		// The Bridge API saves an additional node mask for the virtual nodes in the
 		// previous module, so we can use this offset to distinguish between the two.
-		offset: 0x2fde,
+		offset: 0x2fc1,
 		count: 1,
 	},
 	{ name: "EEOFFSET_MAGIC_far", type: NVMEntryType.Byte, count: 1 },
@@ -182,9 +177,9 @@ const NVM_Layout_Bridge_6_6x: NVMLayout = [
 	{ name: "nvmModuleSizeEndMarker", type: NVMEntryType.Word, count: 1 },
 ];
 
-export const Bridge_6_6x: NVM500Details = {
-	name: "Bridge 6.6x",
-	library: "bridge",
+export const Static_6_6x: NVM500Impl = {
+	name: "Static 6.6x",
+	library: "static",
 	protocolVersions: ["4.33", "4.62"],
-	layout: NVM_Layout_Bridge_6_6x,
+	layout: NVM_Layout_Static_6_6x,
 };
