+++ packages/nvmedit/src/lib/nvm500/impls/Static_6_8x.ts
@@ -1,6 +1,6 @@
 import { MAX_NODES, NUM_NODEMASK_BYTES } from "@zwave-js/core/safe";
 import { SUC_MAX_UPDATES } from "../../../consts";
-import type { NVM500Details } from "../shared";
+import type { NVM500Impl } from "../shared";
 import {
 	APPL_NODEPARM_MAX,
 	NVMEntryType,
@@ -12,7 +12,7 @@ import {
 	TOTAL_RTC_TIMER_MAX,
 } from "../shared";
 
-const NVM_Layout_Static_6_7x: NVMLayout = [
+const NVM_Layout_Static_6_8x: NVMLayout = [
 	{ name: "nvmTotalEnd", type: NVMEntryType.Word, count: 1 },
 	{ name: "nvmZWlibrarySize", type: NVMEntryType.NVMModuleSize, count: 1 },
 	{ name: "NVM_INTERNAL_RESERVED_1_far", type: NVMEntryType.Byte, count: 4 },
@@ -109,6 +109,7 @@ const NVM_Layout_Static_6_7x: NVMLayout = [
 	{
 		name: "EX_NVM_ROUTECACHE_START_far",
 		type: NVMEntryType.Route,
+		offset: 0x268e,
 		count: MAX_NODES,
 	},
 	{
@@ -128,6 +129,7 @@ const NVM_Layout_Static_6_7x: NVMLayout = [
 		size: 16,
 		count: 1,
 	},
+	{ name: "NVM_SYSTEM_STATE", type: NVMEntryType.Byte, count: 1 },
 	{
 		name: "nvmZWlibraryDescriptor",
 		type: NVMEntryType.NVMModuleDescriptor,
@@ -138,7 +140,7 @@ const NVM_Layout_Static_6_7x: NVMLayout = [
 		type: NVMEntryType.NVMModuleSize,
 		// The Bridge API saves an additional node mask for the virtual nodes in the
 		// previous module, so we can use this offset to distinguish between the two.
-		offset: 0x2fd1,
+		offset: 0x2fd2,
 		count: 1,
 	},
 	{ name: "EEOFFSET_MAGIC_far", type: NVMEntryType.Byte, count: 1 },
@@ -162,6 +164,21 @@ const NVM_Layout_Static_6_7x: NVMLayout = [
 		name: "EEOFFSET_POWERLEVEL_LOW_far",
 		type: NVMEntryType.Byte,
 		count: POWERLEVEL_CHANNELS,
+	},
+	{
+		name: "EEOFFSET_MODULE_POWER_MODE_EXTINT_ENABLE_far",
+		type: NVMEntryType.Byte,
+		count: 1,
+	},
+	{
+		name: "EEOFFSET_MODULE_POWER_MODE_far",
+		type: NVMEntryType.Byte,
+		count: 1,
+	},
+	{
+		name: "EEOFFSET_MODULE_POWER_MODE_WUT_TIMEOUT_far",
+		type: NVMEntryType.DWord,
+		count: 1,
 	},
 	{
 		name: "nvmApplicationDescriptor",
@@ -194,9 +211,20 @@ const NVM_Layout_Static_6_7x: NVMLayout = [
 	{ name: "nvmModuleSizeEndMarker", type: NVMEntryType.Word, count: 1 },
 ];
 
-export const Static_6_7x: NVM500Details = {
-	name: "Static 6.7x",
+export const Static_6_8x: NVM500Impl = {
+	name: "Static 6.8x",
 	library: "static",
-	protocolVersions: ["4.60", "4.61", "5.02", "5.03"],
-	layout: NVM_Layout_Static_6_7x,
+	protocolVersions: [
+		"6.01",
+		"6.02",
+		"6.03",
+		"6.04",
+		"6.05",
+		"6.06",
+		"6.07",
+		"6.08",
+		"6.09",
+		"6.10",
+	],
+	layout: NVM_Layout_Static_6_8x,
 };
